datapath: remove 'unparallel' build tag in tests
We've gone through great lengths to avoid using build tags in tests since it
messes with editor tooling, code completion, go-to reference, and notably
excludes files from golangci-lint.

Remove the 'unparallel' build tag.
